requirements.txt
playwright
python-dotenv
fastapi
uvicorn
apscheduler
playwright-stealth
